refactor(tracking-report): 统一抽离类型定义到types目录
将分散在各个文件中的报告生成相关类型统一抽离到src/tracking-report/types目录下，移除重复的类型定义，优化代码结构与维护性
