+++ src/tracking-report/types/index.ts
@@ -1,0 +1,2 @@
+export * from './domain.types';
+export * from './job.types';
